@@ -1,33 +1,30 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.OpenApi.Models;
 
-namespace DaprDemo.Api
+namespace DaprDemo.Api;
+
+public class Startup
 {
-    public class Startup
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public void ConfigureServices(IServiceCollection services)
     {
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddScoped<IHandler<int, int>, Handlers.Slow.Handler>();
-            services.AddControllers();
-            services.AddSwaggerGen(c => { c.SwaggerDoc("v1", new OpenApiInfo {Title = "DaprDemo", Version = "v1"}); });
-        }
+        services.AddScoped<IHandler<int, int>, Handlers.Slow.Handler>();
+        services.AddControllers();
+    }
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            app.UseHttpsRedirection();
-            app.UseRouting();
-            app.UseAuthorization();
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        app.UseHttpsRedirection();
+        app.UseRouting();
+        app.UseAuthorization();
 
-            app.UseCloudEvents();
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-                endpoints.MapSubscribeHandler();
-            });
-        }
+        app.UseCloudEvents();
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllers();
+            endpoints.MapSubscribeHandler();
+        });
     }
 }
